webapp/requirements.txt
openai==0.28.0
python-dotenv
streamlit
streamlit-chat
tiktoken
requests
pillow
